# src/main/java/org/Autogenerated/src/main/java/io/codejournal/maven/swagger2java/api/AuthorApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.28).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.codejournal.maven.swagger2java.api;

import io.codejournal.maven.swagger2java.model.Author;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2022-09-09T15:04:09.579+05:30")

@Validated
@Api(value = "Author", description = "the Author API")
@RequestMapping(value = "/v2")
public interface AuthorApi {

    @ApiOperation(value = "Add a new Author object", nickname = "addAuthor", notes = "Add a new Author object value.", tags={ "Author", })
    @ApiResponses(value = { 
        @ApiResponse(code = 405, message = "Invalid input") })
    @RequestMapping(value = "/Author",
        produces = { "application/xml", "application/json" }, 
        consumes = { "application/json", "application/xml" },
        method = RequestMethod.POST)
    ResponseEntity<Void> addAuthor(@ApiParam(value = "Pet object that needs to be added to the store" ,required=true )  @Valid @RequestBody Author body);


    @ApiOperation(value = "Deletes a Author", nickname = "deleteAuthor", notes = "", tags={ "Author", })
    @ApiResponses(value = { 
        @ApiResponse(code = 400, message = "Invalid Author value") })
    @RequestMapping(value = "/Author/{ID}",
        produces = { "application/xml", "application/json" }, 
        method = RequestMethod.DELETE)
    ResponseEntity<Void> deleteAuthor(@ApiParam(value = "Pet Author to delete",required=true) @PathVariable("ID") Long ID);


    @ApiOperation(value = "Find Author by ID", nickname = "getAuthorByID", notes = "Returns a single Author", response = Author.class, tags={ "Author", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successful operation", response = Author.class),
        @ApiResponse(code = 400, message = "Invalid ID supplied"),
        @ApiResponse(code = 404, message = "Author not found") })
    @RequestMapping(value = "/Author/{ID}",
        produces = { "application/xml", "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<Author> getAuthorByID(@ApiParam(value = "ID of Author to return",required=true) @PathVariable("ID") Long ID);

}